fix some authorization bugs

+++ App/BackEnd/App.Content.Domain/Operations/Commands/DeleteArticle/DeleteArticleCommandHandler.cs
@@ -27,6 +27,10 @@ namespace App.Content.Domain.Operations.Commands.DeleteArticle
             if ( !await _articleRepository.ExistsBySlug(request.Slug) )
                 return OperationResponseFactory.NotFound<DeleteArticleCommand, OperationResponse<DeleteArticleCommandResult>>(typeof(ArticleEntity), request.Slug);
 
+            var article = await _articleRepository.GetBySlug(request.Slug, _userContext.UserId);
+            if (article.Author.UserId != _userContext.UserId)
+                return OperationResponseFactory.NotAuthorized<DeleteArticleCommand, OperationResponse<DeleteArticleCommandResult>>("article does not belong to the user");
+            
             await _articleRepository.Delete(_userContext.UserId, request.Slug);
 
             return OperationResponseFactory.Success(new DeleteArticleCommandResult());
